requirements.txt
flask>=3.1.0
fastapi>=0.115.8
uvicorn>=0.34.0
requests>=2.32.3
pillow>=11.1.0
python-multipart>=0.0.20
numpy>=2.2.2
matplotlib>=3.10.0
